Navigation between new journal pages complete
The navigation between the new structure for the journaling pages is complete.

Small error currently exists where editing an entry, and clicking on save changes, is that when you return to the view entry screen, the changes do not appear there.

The changes are being saved, as they will appear once navigating back to the main Journaling screen.

Debugging will be done in next step